Add files via upload
I made this as my A Level computer science coursework and scored 71/75. Some instances of recursion etc. could most likely be replaced with iteration for a more efficient time complexity but I used as many varying techniques for the sake of marks. The program demonstrates a simulation of a chemical reaction in the most abstract form as a way to demonstrate the effects of various conditions on a reaction or dynamic equilibrium